# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: StatusPage API
  description: API for managing service statuses and incidents
  version: 2.15.0
  contact:
    name: API Support
servers:
  - url: /
    description: Relative paths (for testing)
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/services:
    get:
      tags: [services]
      summary: List services
      description: Public endpoint, no authentication required
      operationId: listServices
      parameters:
        - name: group_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by group membership
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ServiceStatus'
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
          description: Include archived services in the response
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesResponse'
    post:
      tags: [services]
      summary: Create a service
      operationId: createService
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/services/{slug}:
    get:
      tags: [services]
      summary: Get a service by slug
      operationId: getService
      parameters:
        - $ref: '#/components/parameters/ServiceSlug'
      responses:
        '200':
          description: Service data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [services]
      summary: Update a service
      operationId: updateService
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: Service updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [services]
      summary: Archive a service (soft delete)
      description: Archives the service. Returns 409 if service has active events (not resolved, completed, or scheduled).
      operationId: deleteService
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceSlug'
      responses:
        '204':
          description: Service archived
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /api/v1/groups:
    get:
      tags: [groups]
      summary: List groups
      operationId: listGroups
      parameters:
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
          description: Include archived groups in the response
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
    post:
      tags: [groups]
      summary: Create a group
      operationId: createGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/v1/groups/{slug}:
    get:
      tags: [groups]
      summary: Get a group by slug
      operationId: getGroup
      parameters:
        - $ref: '#/components/parameters/GroupSlug'
      responses:
        '200':
          description: Group data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [groups]
      summary: Update a group
      operationId: updateGroup
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupSlug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [groups]
      summary: Archive a group (soft delete)
      description: Archives the group. Returns 409 if group has active events (not resolved, completed, or scheduled) or has non-archived services assigned.
      operationId: deleteGroup
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupSlug'
      responses:
        '204':
          description: Group archived
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        For API clients (curl, Postman, integrations).
        Pass the access token in the Authorization header.
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        For browser-based clients.
        Cookies are automatically set by /auth/login and sent by the browser.
        State-changing requests (POST, PUT, PATCH, DELETE) require X-CSRF-Token header
        matching the csrf_token cookie value.
  parameters:
    ServiceSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
    GroupSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  details:
                    oneOf:
                      - type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                            message:
                              type: string
                          required: [field, message]
                      - type: string
                    description: Validation error details. Array of field errors or error string.
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
  schemas:
    ServiceStatus:
      type: string
      enum: [operational, degraded, partial_outage, major_outage, maintenance]
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/ServiceStatus'
          description: Stored status (set manually or by closed events)
        effective_status:
          $ref: '#/components/schemas/ServiceStatus'
          description: Effective status (worst-case from active events, or stored if no active events). Active events exclude resolved, completed, and scheduled.
        has_active_events:
          type: boolean
          description: Whether the service has any active events (not resolved, completed, or scheduled)
        group_ids:
          type: array
          items:
            type: string
            format: uuid
        order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
          nullable: true
      required: [id, name, slug, status, effective_status, has_active_events, group_ids, order, created_at, updated_at]
    ServiceGroup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        service_ids:
          type: array
          items:
            type: string
            format: uuid
        order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
          nullable: true
      required: [id, name, slug, service_ids, order, created_at, updated_at]
...
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        status:
          $ref: '#/components/schemas/ServiceStatus'
        group_ids:
          type: array
          items:
            type: string
            format: uuid
        order:
          type: integer
          default: 0
        tags:
          type: object
          additionalProperties:
            type: string
      required: [name, slug]
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        status:
          $ref: '#/components/schemas/ServiceStatus'
        group_ids:
          type: array
          items:
            type: string
            format: uuid
        order:
          type: integer
        reason:
          type: string
          description: Reason for status change (recorded in audit log)
      required: [name, slug, status]
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        order:
          type: integer
          default: 0
      required: [name, slug]
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        slug:
          type: string
          minLength: 1
          maxLength: 255
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        description:
          type: string
        order:
          type: integer
        service_ids:
          type: array
          items:
            type: string
            format: uuid
          description: List of service IDs to assign to this group. If provided, replaces all existing service memberships.
      required: [name, slug]
...
    ServiceResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Service'
    ServicesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Service'
    GroupResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ServiceGroup'
    GroupsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ServiceGroup'
...